django_cache/requirements.txt
Django==4.2.1
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
